+++ app/helpers/project_helper.rb
@@ -16,7 +16,11 @@ module ProjectHelper
   end
 
   def converting_on_date(project)
-    date = project.provisional_conversion_date.to_formatted_s(:govuk)
+    conversion_date = project.conversion_date
+
+    return conversion_date.date.to_formatted_s(:govuk) unless conversion_date.provisional?
+
+    date = conversion_date.date.to_formatted_s(:govuk)
     tag = govuk_tag(text: "provisional", colour: "grey")
 
     "#{date} #{tag}".html_safe