+++ spec/helpers/project_helper_spec.rb
@@ -56,11 +56,23 @@ RSpec.describe ProjectHelper, type: :helper do
   end
 
   describe "#converting_on_date" do
-    it "returns the formatted date and a provisional tag" do
-      project = build(:conversion_project)
+    context "when the conversion date is provisional" do
+      it "returns the formatted date and a provisional tag" do
+        project = build(:conversion_project)
 
-      expect(helper.converting_on_date(project)).to include project.provisional_conversion_date.to_formatted_s(:govuk)
-      expect(helper.converting_on_date(project)).to include "provisional"
+        expect(helper.converting_on_date(project)).to include project.provisional_conversion_date.to_formatted_s(:govuk)
+        expect(helper.converting_on_date(project)).to include "provisional"
+      end
+    end
+
+    context "when the conversion date is confirmed" do
+      it "returns the formatted date and a provisional tag" do
+        project = build(:conversion_project)
+        project.task_list.stakeholder_kick_off_confirmed_conversion_date = Date.today
+
+        expect(helper.converting_on_date(project)).to include project.task_list.stakeholder_kick_off_confirmed_conversion_date.to_formatted_s(:govuk)
+        expect(helper.converting_on_date(project)).not_to include "provisional"
+      end
     end
   end
 end