+++ spec/models/conversion/project_spec.rb
@@ -20,4 +20,40 @@ RSpec.describe Conversion::Project do
       end
     end
   end
+
+  describe "#conversion_date" do
+    context "when the project is an involuntary conversion" do
+      let(:project) { build(:involuntary_conversion_project) }
+
+      it "the conversion date is always provisional" do
+        conversion_date = project.conversion_date
+
+        expect(conversion_date.date).to eq project.provisional_conversion_date
+        expect(conversion_date.provisional?).to eq true
+      end
+    end
+
+    context "when the project has no confirmed conversion date" do
+      let(:project) { build(:voluntary_conversion_project) }
+
+      it "returns the appropriate values" do
+        conversion_date = project.conversion_date
+
+        expect(conversion_date.date).to eq project.provisional_conversion_date
+        expect(conversion_date.provisional?).to eq true
+      end
+    end
+
+    context "when the project has a confirmed conversion date from the external stakeholder kick off task" do
+      let(:project) { build(:voluntary_conversion_project) }
+
+      it "returns the appropriate values" do
+        project.task_list.stakeholder_kick_off_confirmed_conversion_date = Date.today + 1.year
+        conversion_date = project.conversion_date
+
+        expect(conversion_date.date).to eq project.task_list.stakeholder_kick_off_confirmed_conversion_date
+        expect(conversion_date.provisional?).to eq false
+      end
+    end
+  end
 end